Добавлена независимость Header компонента & изменение структуры проекта
