--- common/containers/index.ts
@@ -1,2 +1,0 @@
-export { Header } from "./Header";
-export { HomeContent } from "./HomeContent";
