+++ containers/index.ts
@@ -1,0 +1,2 @@
+export { Header } from "../common/lib/Header";
+export { HomeContent } from "./HomeContent";
